4a-docs/servUsuarios/sup_be/authApp/serializers/userSerializer.py
from rest_framework import fields, serializers
from authApp.models.user import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields= ['id', 'username', 'password', 'name', 'email' ]
        